requirements.txt
opencv-python==4.7.0.72
cmake==3.26.3
dlib==19.24.1
pillow==9.5.0
